debug counters made double
git-svn-id: svn+ssh://fwe02.ifj.edu.pl/xdat1/lhcb/mchrzaszcz/svn/KK-all-2012@978 957d80b3-de0b-4737-8e94-343ebb428fb4

+++ MaMar/TMCgenFOAM.h
@@ -143,7 +143,7 @@ class TMCgenFOAM :public TMCgen
   double m_p4[4];         //!
   // Model Weights
   double m_WTmodel[100];  //!
-  long   m_count;
+  double m_count;
 ///////////////////////////////////////////////////////////
 /// methods obligatory
   void Initialize(TRandom*, ofstream*, TH1D*);
